added all test functions to Test class
added all the functions tht work correctly for testing with LinkedList functions.
