+++ CustomeDB/CustomeDB.cpp
@@ -1,25 +1,23 @@
 #include <iostream>
 #include <string>
 #include "LinkedList.h"
-#include "Test/Test.h"
+#include "Test.h"
 
 using namespace std;
 
 
-
-
 int main()
 {
-    LinkedList l1("myFile.txt");
-    /*l1.addAtEnd(1, "talha", 20);
-    l1.addAtEnd(2, "haseed", 10);
-    l1.addAtEnd(3, "bats", 25);*/
-    l1.displayList();
-    LinkedList l2("newFile.txt");
-    /*l2.addAtEnd(4, "Abad", 30);*/
-    l2.displayList();
-
-    Test obj();
+    //LinkedList l1("myFile.txt");
+    ///*l1.addAtEnd(1, "talha", 20);
+    //l1.addAtEnd(2, "haseed", 10);
+    //l1.addAtEnd(3, "bats", 25);*/
+    //l1.displayList();
+    //LinkedList l2("newFile.txt");
+    ///*l2.addAtEnd(4, "Abad", 30);*/
+    //l2.displayList();
+    cout << 'g' << endl;
+    Test ();
 
     return 0;
 }
